+++ R/hbm-info.R
@@ -33,7 +33,7 @@ NULL
 #' data("data_fhnorm")
 #' model <- hbm(brms::bf(y ~ x1), data = data_fhnorm,
 #'              re = ~ (1 | regency),    # area-level random effect
-#'              chains = 2, iter = 1000, warmup = 500,
+#'              chains = 4, iter = 2000, warmup = 1000,
 #'              cores = 1, seed = 1, refresh = 0)
 #' hbm_info(model)
 #' }
@@ -82,7 +82,7 @@ hbm_info <- function(model) {
 #' data("data_fhnorm")
 #' model <- hbm(brms::bf(y ~ x1), data = data_fhnorm,
 #'              re = ~ (1 | regency),    # area-level random effect
-#'              chains = 2, iter = 1000, warmup = 500,
+#'              chains = 4, iter = 2000, warmup = 1000,
 #'              cores = 1, seed = 1, refresh = 0)
 #' head(hbm_data(model))
 #' }
@@ -114,7 +114,7 @@ hbm_data <- function(model) {
 #' data("data_fhnorm")
 #' model <- hbm(brms::bf(y ~ x1), data = data_fhnorm,
 #'              re = ~ (1 | regency),    # area-level random effect
-#'              chains = 2, iter = 1000, warmup = 500,
+#'              chains = 4, iter = 2000, warmup = 1000,
 #'              cores = 1, seed = 1, refresh = 0)
 #' hbm_warnings(model)
 #' }