+++ R/hbm_binlogitnorm.R
@@ -86,7 +86,7 @@
 #'   auxiliary  = c("x1", "x2", "x3"),
 #'   area_var   = "district",        # area-level random effect (1 | district)
 #'   data       = data_binlogitnorm,
-#'   chains = 1, iter = 500, warmup = 250, refresh = 0
+#'   chains = 4, iter = 2000, warmup = 1000, refresh = 0
 #' )
 #'
 #' # -- 2. With spatial CAR random effect -------------------------------------
@@ -99,7 +99,7 @@
 #'   spatial_model = "car",
 #'   M             = adjacency_matrix_car_regency,
 #'   data          = data_binlogitnorm,
-#'   chains = 1, iter = 500, warmup = 250, refresh = 0
+#'   chains = 4, iter = 2000, warmup = 1000, refresh = 0
 #' )
 #' }
 #'